requirements.txt
python-dotenv
fastapi
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
uvicorn
Werkzeug
pyjwt